refactoring fixed "off-by-one" error
