+++ internal/persistence/persistence_test.go
@@ -3,6 +3,7 @@ package persistence
 import (
 	"github.com/markusressel/fan2go/internal/configuration"
 	"github.com/markusressel/fan2go/internal/fans"
+	"github.com/markusressel/fan2go/internal/util"
 	"github.com/stretchr/testify/assert"
 	"testing"
 )
@@ -42,9 +43,10 @@ func TestReadFan(t *testing.T) {
 	persistence := NewPersistence(dbTestingPath)
 
 	fan, _ := createFan(false, NeverStoppingFan)
-	err := persistence.SaveFanPwmData(fan)
+	expected := util.InterpolateLinearly(fan.GetFanCurveData(), 0, 256)
 
-	fan, _ = createFan(false, LinearFan)
+	err := persistence.SaveFanPwmData(fan)
+	assert.NoError(t, err)
 
 	// WHEN
 	fanData, err := persistence.LoadFanPwmData(fan)
@@ -52,6 +54,7 @@ func TestReadFan(t *testing.T) {
 	// THEN
 	assert.Nil(t, err)
 	assert.NotNil(t, fanData)
+	assert.Equal(t, expected, fanData)
 }
 
 func createFan(neverStop bool, curveData map[int]float64) (fan fans.Fan, err error) {
